Run the IRate Management Procedure

Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque eleifend odio ac rutrum luctus. Aenean placerat porttitor commodo. Pellentesque eget porta libero. Pellentesque molestie mi sed orci feugiat, non mollis enim tristique. Suspendisse eu sapien vitae arcu lobortis ultrices vitae ac velit. Curabitur id nunc euismod ante fringilla lobortis. Aliquam ullamcorper in diam non placerat.

IRate(omp, sr, years, yref = seq(dims(omp)$minyear, years[1] - 1),
  responsiveness, hr_multiplier, biomass_threshold, biomass_limit, maxTAC,
  DLAG = 1, MLAG = 1, SFREQ = 1, errcpue = ~0, effcpue = ~0,
  srresiduals = ~0, errimp = ~0, verbose = FALSE)

Arguments

omp

An FLStock object with the extended OM

sr

A list with SR params and model, to be passed to fwd()

years

For which years is the simulation to be run?

verbose

Should progress be reported on screen? Defaults to TRUE

Value

FLStock

See also

FLStock

Examples

data(om)
omp <- fwdWindow(om, end=dims(om)$maxyear + 3, br)

#> Error in fwdWindow(om, end = dims(om)$maxyear + 3, br): object 'br' not found

# Hindcast run
run <- IRate(omp, sro, years=1990:2012, yref=c(1981:2010),
  responsiveness=0.5, DLAG=1, MLAG=1, SFREQ=2,
  hr_multiplier= 1.1, biomass_threshold=0.5,
  biomass_limit=0.2, maxTAC=500000,
  srresiduals=~rlnorm(dims(om)$iter, 0, 0.5), verbose=FALSE)

#> Error in sel(omp): object 'omp' not found

plot(window(FLStocks(run=run, om=om), end=2012))

#> Error in FLStocks(run = run, om = om): object 'run' not found